Interactive map, updated index  and main.css page
I created an interactive map of police shootings in the US using the
leaflet API and instructions provided on course webpage. The index.html
page was updated with the appropriate elements. I added minimal
stylization to the main.css page. Added in a server.py folder in order
to run a local server on my computer.